requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-groq
python-dotenv
pdfplumber
yagmail
neo4j